[DEBUG] Packet is not forwarded to controller
Even if the transport is destroyed and the connection does not
exist in uid_map_, the packet is not forwarded to controller.
This error occurs when destination address exists in peerlist in tap
module and forwarded to tincan manager with non-zero destination uid.
In detail, after connection is destroyed, we still keep the ipv4-uid
mapping in tap peerlist and forward to tincan manager with its
previous uid. So I add logic in tincan manager to forward packet with
non-zero uid that does not exist in uid_map_.