Add tincanxmppsocket to enable DTLS on WIN32
tincanxmppsocket.cc is also exactly the same as xmppsocket.cc and
it was the only way I could think of to enable DTLS on WIN32.
I attempted creating a subclass of XmppSocket but that did not
work because the CreateCricketSocket is private instead of
protected (along with necessary state). So the least intrusive
solution was to copy and paste xmppsocket.cc and modify locally